:truck: refactor(api) move repo to configuration context
